requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.6
django-guardian==2.4.0
django-import-export==3.3.1
docopt==0.6.2
et-xmlfile==1.1.0
idna==3.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
Pillow==10.1.0
PyYAML==6.0.1
psycopg2-binary==2.9.9
requests==2.31.0
setuptools==68.0.0
sqlparse==0.4.4
tablib==3.5.0
urllib3==2.0.7
wheel==0.37.1
xlrd==2.0.1
xlwt==1.3.0
yarg==0.1.9